+++ src/programs/voice/voice-host.ts
@@ -1,5 +1,7 @@
-import { Message } from "discord.js";
-import VoiceOnDemandTools from "./common";
+import { Message, VoiceState } from "discord.js";
+import VoiceOnDemandTools, {
+  voiceOnDemandRequestHostIdentifier,
+} from "./common";
 import { hasRole } from "../../common/moderator";
 import Tools from "../../common/tools";
 import {
@@ -8,6 +10,8 @@ import {
   CommandHandler,
 } from "../../event-distribution";
 import prisma from "../../prisma";
+import { VoiceStateChange } from "../../event-distribution/events/voice-state-update";
+import { TimerService } from "../timer/timer.service";
 
 @Command({
   event: DiscordEvent.MESSAGE,
@@ -87,6 +91,38 @@ class ChangeHostOnDemand implements CommandHandler<DiscordEvent.MESSAGE> {
 
     await message.reply(
       `I transfered ownership of your room to <@${mentionedMember.id}>!`
+    );
+  }
+}
+
+@Command({
+  event: DiscordEvent.VOICE_STATE_UPDATE,
+  changes: [VoiceStateChange.LEFT, VoiceStateChange.SWITCHED_CHANNEL],
+})
+class RequestNewHostIfNeeded
+  implements CommandHandler<DiscordEvent.VOICE_STATE_UPDATE>
+{
+  async handle(oldState: VoiceState, newState: VoiceState): Promise<void> {
+    if (oldState.channelID === newState.channelID) return;
+
+    const channelId = oldState.channel.id;
+    const mapping = await prisma.voiceOnDemandMapping.findUnique({
+      where: { channelId },
+    });
+
+    if (!mapping) return;
+    //We don't care about any other user else than the host leaving, this should avoid spamming the DB
+    if (oldState.member.id !== mapping.userId) return;
+
+    const executeTime = new Date();
+    executeTime.setMinutes(executeTime.getMinutes() + 1);
+    executeTime.setSeconds(executeTime.getSeconds() + 1);
+    await TimerService.createTimer(
+      voiceOnDemandRequestHostIdentifier,
+      executeTime,
+      {
+        channelId: channelId,
+      }
     );
   }
 }
